src/main/java/org/automation/pages/LoginPage.java
package org.automation.pages;

import org.openqa.selenium.WebDriver;
import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

public class LoginPage {
    private WebDriver driver;

    @FindBy(css = "span.base")
    private WebElement customerLogInText;

    public LoginPage(WebDriver driver) {
        this.driver = driver;
        PageFactory.initElements(driver, this);
    }

    public String getLoginPageTitle() {
        return driver.getTitle();
    }

    public String getCustomerLoginText() {
        return customerLogInText.getText();
    }
}
